+++ lib/data/model/response/setting_model.dart
@@ -1,3 +1,5 @@
+import 'dart:io';
+
 class SettingModel {
   String termsAndConditions;
   String privacyPolicy;
@@ -19,7 +21,10 @@ class SettingModel {
         privacyPolicy: json["privacy_policy"]["value"],
         userAgreement: json["user_agreement"]["value"],
         cancelAnytime: json["cancel_anytime"]["value"],
-        freeGenerations: json["free_generations"] ?? 0,
+        freeGenerations: (Platform.isAndroid
+                ? json["free_generations"]
+                : json["ios_free_generations"]) ??
+            0,
       );
 
   // to json