+++ lib/data/service/generation_service.dart
@@ -1,5 +1,5 @@
 abstract class GenerationServiceInterface {
   Future<int> loadDailyGenerationCount();
   Future<void> incrementDailyGenerationCount();
-  bool canGenerateImage();
+  Future<void> resetDailyGenerationCount();
 }