+++ lib/data/service/generation_service_interface.dart
@@ -1,8 +1,6 @@
 import 'package:intl/intl.dart';
-import 'package:matrix_ai/controller/generation_controller.dart';
 import 'package:matrix_ai/data/repository/generation_repo_interface.dart';
 import 'package:matrix_ai/data/service/generation_service.dart';
-import '../../controller/settings_controller.dart';
 
 class GenerationService implements GenerationServiceInterface {
   final GenerationRepoInterface generationRepoInterface;
@@ -30,12 +28,8 @@ class GenerationService implements GenerationServiceInterface {
   }
 
   @override
-  bool canGenerateImage() {
-    if (SettingsController.find.settingModel.freeGenerations > 0) {
-      return GenerationController.find.dailyGenerationCount <
-          SettingsController.find.settingModel.freeGenerations;
-    } else {
-      return true;
-    }
+  Future<void> resetDailyGenerationCount() async {
+    final key = _generateKey();
+    await generationRepoInterface.setDailyGenerationCount(key, 0);
   }
 }